requirements.txt
PyGObject~=3.42.0
google-api-python-client~=2.38.0
google-auth-httplib2~=0.1.0
google-auth-oauthlib~=0.4.6
